+++ addons/dialogic/Other/DialogicClass.gd
@@ -113,24 +113,15 @@ static func load(slot_name: String = ''):
 ## @param slot_name		The name of the save slot. To load this save you have to specify the same
 ##						If the slot folder doesn't exist it will be created. 
 ##						Leaving this empty will use the last loaded save slot.
-static func save(slot_name: String = '', check_autosave = false) -> void:
-	# check if to save
-	if check_autosave and not get_autosave():
+static func save(slot_name: String = '', is_autosave = false) -> void:
+	# check if to save (if this is a autosave)
+	if is_autosave and not get_autosave():
 		return
-	
-	# make sure to save to the current slot
-	if slot_name == '':
-		slot_name = get_current_slot()
-		if slot_name == '/':
-			slot_name = ''
-			print('[D] Seems like you imported data and now called Dialogic.save(). It will be saved to the default slot. If you use import() though, you might want to use export() as well. Check the docs on saving and loading! ')
 	
 	# gather the info
 	var current_dialog_info = {}
 	if has_current_dialog_node():
 		current_dialog_info = Engine.get_main_loop().get_meta('latest_dialogic_node').get_current_state_info()
-	else:
-		print('[D] You called Dialogic.save() but no dialog node was instanced!')
 	
 	var game_state = {}
 	if Engine.get_main_loop().has_meta('game_state'):