+++ addons/dialogic/Documentation/Content/Changelog.md
@@ -22,6 +22,7 @@
 - Updated UI of the AudioPickers [[Jowan-Spooner]](https://github.com/Jowan-Spooner)
 - Automatically scrolling to newly created events when you click on a create event button [[Jowan-Spooner](https://github.com/Jowan-Spooner)]
 - If no characters are present in the current project the character picker is hidden and a prompt was added in the Character join and Character leave events to create one [[zaknafean](https://github.com/zaknafean)]
+- Copy Timeline Name will now return the full path to that timeline
 
 #### Theme Editor:
 - **Updated Audio Settings** [[Tim Krief](https://github.com/timkrief)]: